renamed expand to raml
