+++ lib/cli/__test__/raml.spec.js
@@ -1,12 +1,12 @@
 var _ = require('lodash')
 var BPromise = require('bluebird')
-var buildRAML = require('../expand').buildRAML
-var expand = require('../expand')
+var buildRAML = require('../raml').buildRAML
+var expand = require('../raml').buildJSON
 var expect = require('chai').expect
 var fs = require('fs-extra')
 var path = require('path')
 
-describe('expand()', function () {
+describe('raml()', function () {
     this.timeout(10e3)
 
     it('should throw invalid RAML error', function (done) {
