+++ lib/cli/__test__/transform.spec.js
@@ -1,5 +1,5 @@
 var emptyScalar = require('./utils').emptyScalar
-var expand = require('../expand')
+var expand = require('../raml').buildJSON
 var expect = require('chai').expect
 var get = require('lodash/get')
 var transform = require('../transform')
