+++ lib/cli/build.js
@@ -1,7 +1,7 @@
 var _ = require('lodash')
 var BPromise = require('bluebird')
-var buildRAML = require('./expand').buildRAML
-var expand = require('./expand')
+var buildRAML = require('./raml').buildRAML
+var expand = require('./raml').buildJSON
 var fs = require('fs-extra')
 var Joi = require('joi')
 var Handlebars = require('handlebars')
