+++ lib/cli/raml.js
@@ -18,8 +18,8 @@ BPromise.promisifyAll(fs)
 BPromise.promisifyAll(glob)
 BPromise.promisifyAll(refs)
 
-module.exports = buildJSON
-module.exports.buildRAML = buildRAML
+exports.buildJSON = buildJSON
+exports.buildRAML = buildRAML
 
 function prepareRAML (options) {
     if (!options.tmpDir) options.tmpDir = constants.tmpDir
